added some method

+++ internal/pkg/mark/interface.go
@@ -1,5 +1,5 @@
 package mark
 
 type Marker interface {
-	MarkResult(answer string) (int, error)
+	MarkResult(answer string) (int8, error)
 }
